Correcion de errores e implementacion de sugerencias
