+++ src/app/layout.js
@@ -4,21 +4,21 @@ import { ConnectedRouter } from 'connected-react-router';
 import { useDispatch } from 'react-redux';
 
 import { history } from 'redux/store';
-import { HOME, BILLS_HISTORY, LIST_OF_PAYMENTS } from 'constants/routes';
+import { HOME, BILLS_HISTORY, PAYMENTS_HISTORY } from 'constants/routes';
 import PaymentActions from 'redux/payment/actions';
 
 import Topbar from './components/Topbar';
 import Home from './screens/Home';
 import BillsHistory from './screens/BillsHistory';
 import styles from './styles.module.scss';
-import PaymentHistory from './screens/PaymentHistory';
-import PaymentDetail from './screens/PaymentDetail';
+import PaymentsHistory from './screens/PaymentsHistory';
+import PaymentDetail from './screens/PaymentsHistory/Components/PaymentDetail';
 
 const App = () => {
   const dispatch = useDispatch();
 
   const onChangeRouteSetCurrentPayment = () => {
-    if (!history.location.pathname.includes(LIST_OF_PAYMENTS)) {
+    if (!history.location.pathname.includes(PAYMENTS_HISTORY)) {
       dispatch(PaymentActions.setCurrentPayment(null));
     }
   };
@@ -36,8 +36,8 @@ const App = () => {
           <Switch>
             <Route exact path={HOME} component={Home} />
             <Route exact path={BILLS_HISTORY} component={BillsHistory} />
-            <Route exact path={LIST_OF_PAYMENTS} component={PaymentHistory} />
-            <Route exact path={`${LIST_OF_PAYMENTS}/:id`} component={PaymentDetail} />
+            <Route exact path={PAYMENTS_HISTORY} component={PaymentsHistory} />
+            <Route exact path={`${PAYMENTS_HISTORY}/:id`} component={PaymentDetail} />
             <Route render={() => <Redirect to={HOME} />} />
           </Switch>
         </ConnectedRouter>
