+++ src/app/screens/PaymentsHistory/Components/PaymentDetail/styles.module.scss
