+++ src/app/screens/PaymentsHistory/utils/numberUtils.js
@@ -1,0 +1,1 @@
+export const formatNumber = amount => amount.toFixed(1).replace('.', ',');
