+++ src/constants/routes.js
@@ -1,3 +1,3 @@
 export const HOME = '/';
 export const BILLS_HISTORY = '/historico_de_facturas';
-export const LIST_OF_PAYMENTS = '/historico_de_pagos';
+export const PAYMENTS_HISTORY = '/historico_de_pagos';
